rightOperandPrinter is only used in a single place thus it's not needed to be exported
